+++ python/dalex/dalex/__init__.py
@@ -5,10 +5,18 @@ from .arena.object import Arena
 __version__ = '0.4.1.9000'
 
 __all__ = [
-    "Explainer",
-    "model_explanations",
-    "predict_explanations",
-    "fairness",
+    "Arena",
     "datasets",
-    "Arena"
+    "Explainer",
+    "fairness"
 ]
+
+# specify autocompletion in IPython
+# see comment: https://github.com/ska-telescope/katpoint/commit/ed7e8b9e389ee035073c62c2394975fe71031f88
+# __dir__ docs (Python 3.7!): https://docs.python.org/3.7/library/functions.html#dir 
+
+def __dir__():
+    """IPython tab completion seems to respect this."""
+    return __all__ +  ['__all__', '__builtins__', '__cached__', '__doc__', '__file__',
+                       '__loader__', '__name__', '__package__', '__path__', '__spec__',
+                       '__version__']